Rename basic_plugin into minimal + Adjust example code
